Conflicts:
	www/application/bootstrap.php
	www/application/classes/controller/authoredlabyrinth.php
	www/application/classes/controller/base.php
	www/application/classes/controller/chatmanager.php
	www/application/classes/controller/clustermanager.php
	www/application/classes/controller/countermanager.php
	www/application/classes/controller/elementmanager.php
	www/application/classes/controller/feedbackmanager.php
	www/application/classes/controller/filemanager.php
	www/application/classes/controller/home.php
	www/application/classes/controller/labyrinthmanager.php
	www/application/classes/controller/linkmanager.php
	www/application/classes/controller/mapusermanager.php
	www/application/classes/controller/nodemanager.php
	www/application/classes/controller/patient.php
	www/application/classes/controller/popupmanager.php
	www/application/classes/controller/questionmanager.php
	www/application/classes/controller/renderlabyrinth.php
	www/application/classes/controller/reportmanager.php
	www/application/classes/controller/skinmanager.php
	www/application/classes/controller/usermanager.php
	www/application/classes/controller/visualmanager.php
	www/application/classes/controller/webinarmanager.php
	www/application/classes/crossreferences.php
	www/application/classes/model/labyrinth.php
	www/application/classes/model/leap/map.php
	www/application/classes/model/leap/map/avatar.php
	www/application/classes/model/leap/map/counter/commonrules.php
	www/application/classes/model/leap/map/element.php
	www/application/classes/model/leap/map/node.php
	www/application/classes/model/leap/map/node/reference.php
	www/application/classes/model/leap/map/node/section.php
	www/application/classes/model/leap/map/question.php
	www/application/classes/model/leap/map/user.php
	www/application/classes/model/leap/map/vpd.php
	www/application/classes/model/leap/patient.php
	www/application/classes/model/leap/patient/condition.php
	www/application/classes/model/leap/patient/conditionrelation.php
	www/application/classes/model/leap/patient/rule.php
	www/application/classes/model/leap/patient/sessions.php
	www/application/classes/model/leap/user.php
	www/application/classes/model/leap/user/response.php
	www/application/classes/model/leap/user/session.php
	www/application/classes/model/leap/user/sessiontrace.php
	www/application/classes/model/leap/webinar.php
	www/application/classes/model/leap/webinar/map.php
	www/application/classes/model/visualeditor.php
	www/application/classes/runtimelogic.php
	www/application/views/adminMenu.php
	www/application/views/home.php
	www/application/views/labyrinth/authored.php
	www/application/views/labyrinth/file/global.php
	www/application/views/labyrinth/file/view.php
	www/application/views/labyrinth/global.php
	www/application/views/labyrinth/labyrinthEditorMenu.php
	www/application/views/labyrinth/node/addNode.php
	www/application/views/labyrinth/node/editNode.php
	www/application/views/labyrinth/question/area.php
	www/application/views/labyrinth/question/choice.php
	www/application/views/labyrinth/question/draganddrop.php
	www/application/views/labyrinth/question/global.php
	www/application/views/labyrinth/question/text.php
	www/application/views/labyrinth/question/view.php
	www/application/views/labyrinth/report/report.php
	www/application/views/labyrinth/skin/basic/basic.php
	www/application/views/labyrinth/skin/basic/basic_template.php
	www/application/views/labyrinth/user/view.php
	www/application/views/labyrinth/visual.php
	www/application/views/patient/grid.php
	www/application/views/patient/labyrinth.php
	www/application/views/patient/management.php
	www/application/views/userMenu.php
	www/application/views/usermanager/editUser.php
	www/application/views/webinar/render.php
	www/application/views/webinar/statistic.php
	www/application/views/webinar/view.php
	www/application/views/webinar/webinar.php
	www/css/skin/basic/layout_basic.css
	www/scripts/application.js
	www/scripts/avatar.js
	www/scripts/basic.js
	www/scripts/bootstrap/css/bootstrap.css
	www/scripts/rules-checker.js
	www/scripts/visualeditor/rightPanel.js
	www/scripts/webinar.js
	www/scripts/webinar/progress.js

+++ www/scripts/visualeditor/deleteModal.js
@@ -26,7 +26,7 @@ var DeleteModal = function() {
         
         if('rightPanel' in parameters)
             self.rightPanel = parameters.rightPanel;
-    }
+    };
     
     self.Show = function(mode) {
         dialogMode = mode;
@@ -42,7 +42,7 @@ var DeleteModal = function() {
             
             $modal.modal();
         }
-    }
+    };
     
     self.Hide = function() {
         if($modal != null) {
@@ -52,36 +52,31 @@ var DeleteModal = function() {
             $('.deleteModalContentNodes').hide();
             $modal.modal('hide');
         }
-    }
+    };
      
     var ApplyEvent = function() {
         if(dialogMode == 'single') {
             if(self.node != null && visualEditor != null) {
                 visualEditor.DeleteNodeById(self.node.id);
                 self.node = null;
-
                 visualEditor.Render();
             }
         } else if(dialogMode == 'multiple') {
             if(self.selectedNodes != null && self.selectedNodes.length > 0 && visualEditor != null) {
                 for(var i = 0; i < self.selectedNodes.length; i++) {
-                    if(!self.selectedNodes[i].isRoot)
-                    visualEditor.DeleteNodeById(self.selectedNodes[i].id);
+                    if( ! self.selectedNodes[i].isRoot) visualEditor.DeleteNodeById(self.selectedNodes[i].id);
                 }
-                
                 visualEditor.Render();
             }
         }
         
-        if(self.rightPanel != null)
-            self.rightPanel.Hide();
+        if(self.rightPanel != null) self.rightPanel.Hide();
         
         self.Hide();
-        
-        if(self.selectRoot) {
-            utils.ShowMessage($('#ve_message'), $('#ve_message_text'), 'error', 'You cannot delete the root node.', 3000, $('#ve_actionButton'), false);
-        }
+        $('#update').prop('disabled', false).css('background-color', '#777676');
+
+        if(self.selectRoot) utils.ShowMessage($('#ve_message'), $('#ve_message_text'), 'error', 'You cannot delete the root node.', 3000, $('#ve_actionButton'), false);
 
         return false;
     }
-}
+};
